+++ src/app/main/services/component-communication-event.service.ts
@@ -38,5 +38,6 @@ export class EmitEvent {
 export enum Events {
     loginSuccess,
     gravSearchExecuted,
-    projectCreated
+    projectCreated,
+    resourceDeleted
 }