+++ src/app/project/ontology/property-info/property-info.component.ts
@@ -214,7 +214,8 @@ export class PropertyInfoComponent implements OnChanges, AfterContentInit {
                             this.propAttribute = onto.classes[this.propDef.objectType].label;
                             this.propAttributeComment = onto.classes[this.propDef.objectType].comment;
                         }
-                    }
+                    },
+                    () => {} // don't log error to rollbar if 'currentProjectOntologies' does not exist in the cache
                 );
             } else {
                 this.propAttribute = this.ontology.classes[this.propDef.objectType].label;
@@ -256,7 +257,8 @@ export class PropertyInfoComponent implements OnChanges, AfterContentInit {
                             }
                         });
                     });
-                }
+                },
+                () => {} // don't log error to rollbar if 'currentProjectOntologies' does not exist in the cache
             );
         }
     }