+++ src/app/workspace/resource/values/int-value/int-value.component.ts
@@ -6,6 +6,8 @@ import { CustomRegex } from '../custom-regex';
 import { ValueErrorStateMatcher } from '../value-error-state-matcher';
 
 
+const DECIMAL_VALUE = 10;
+
 @Component({
     selector: 'app-int-value',
     templateUrl: './int-value.component.html',
@@ -51,7 +53,7 @@ export class IntValueComponent extends BaseValueDirective implements OnInit, OnC
 
         const newIntValue = new CreateIntValue();
 
-        newIntValue.int = this.valueFormControl.value;
+        newIntValue.int = parseInt(this.valueFormControl.value, DECIMAL_VALUE);
 
         if (this.commentFormControl.value) {
             newIntValue.valueHasComment = this.commentFormControl.value;
@@ -69,7 +71,7 @@ export class IntValueComponent extends BaseValueDirective implements OnInit, OnC
 
         updatedIntValue.id = this.displayValue.id;
 
-        updatedIntValue.int = this.valueFormControl.value;
+        updatedIntValue.int = parseInt(this.valueFormControl.value, DECIMAL_VALUE);
 
         // add the submitted comment to updatedIntValue only if user has added a comment
         if (this.commentFormControl.value) {