+++ src/app/workspace/resource/values/list-value/list-value.component.ts
@@ -14,6 +14,7 @@ import { DspApiConnectionToken } from 'src/app/main/declarations/dsp-api-tokens'
 import { BaseValueDirective } from 'src/app/main/directive/base-value.directive';
 import { ErrorHandlerService } from 'src/app/main/services/error-handler.service';
 
+
 @Component({
     selector: 'app-list-value',
     templateUrl: './list-value.component.html',
@@ -32,6 +33,8 @@ export class ListValueComponent extends BaseValueDirective implements OnInit, On
 
     customValidators = [];
 
+    selectedNodeHierarchy: string[] = [];
+
     constructor(
         @Inject(FormBuilder) protected _fb: FormBuilder,
         @Inject(DspApiConnectionToken) private _dspApiConnection: KnoraApiConnection,
@@ -42,6 +45,7 @@ export class ListValueComponent extends BaseValueDirective implements OnInit, On
 
     getInitValue(): string | null {
         if (this.displayValue !== undefined) {
+            this.getReadModeValue(this.displayValue.listNode);
             return this.displayValue.listNode;
         } else {
             return null;
@@ -126,5 +130,40 @@ export class ListValueComponent extends BaseValueDirective implements OnInit, On
         this.valueFormControl.markAsDirty();
         this.selectedNode = item;
         this.valueFormControl.setValue(item.id);
+    }
+
+    getReadModeValue(nodeIri: string): void {
+        const rootNodeIris = this.propertyDef.guiAttributes;
+        for (const rootNodeIri of rootNodeIris) {
+            const trimmedRootNodeIRI = rootNodeIri.substr(7, rootNodeIri.length - (1 + 7));
+            this._dspApiConnection.v2.list.getList(trimmedRootNodeIRI).subscribe(
+                (response: ListNodeV2) => {
+                    if (!response.children.length) { // this shouldn't happen since users cannot select the root node
+                        this.selectedNodeHierarchy.push(response.label);
+                    } else {
+                        this.selectedNodeHierarchy = this._getHierarchy(nodeIri, response.children);
+                    }
+                }, (error: ApiResponseError) => {
+                    this._errorHandler.showMessage(error);
+                });
+        }
+    }
+
+    _getHierarchy(selectedNodeIri: string, children: ListNodeV2[]): string[] {
+        for (let i = 0; i < children.length; i++) {
+            const node = children[i];
+            if (node.id !== selectedNodeIri) {
+                if (node.children) {
+                    const path = this._getHierarchy(selectedNodeIri, node.children);
+
+                    if (path) {
+                        path.unshift(node.label);
+                        return path;
+                    }
+                }
+            } else {
+                return [node.label];
+            }
+        }
     }
 }